Move the POS refund flow types into their own file
POSOrderListController.swift held five enums that belong to the refund flow
rather than to the order list. They move to POSRefundFlowTypes.swift
unchanged. RefundActionAvailability.unknown is dropped: nothing produced it,
so the order details view no longer branches on it.
